(refactor) some refactoring to improve this
