+++ src/components/icons/CopyIcon.tsx
