+++ src/components/icons/IconDelete.tsx
